Refactor validation logic by moving it to searchValidation.js and updating import paths
