+++ client/src/util/cleanUpText.js
@@ -1,9 +1,9 @@
-export function cleanUpText(text) {
+export function cleanUpText(text = "") {
   return text
     .replace(/\s+/g, " ")
     .replace(/-+/g, "-")
     .replace(/\/+/g, "/")
     .replace(/'+/g, "'")
-    .replace(/.+/g, ".")
+    .replace(/\.+/g, ".")
     .trim();
 }
