+++ client/src/util/searchValidation.js
